server/pinturillo_server/requirements.txt
channels
djangorestframework